llm_module/requirements.txt
# --- Base & API ---
fastapi>=0.115.12,<0.116.0
uvicorn==0.34.3
hypercorn>=0.17.3,<0.18.0    # HTTP/2 (h2c) support pour le controller GAMA
pydantic==2.12.5
pydantic-settings>=2.9.1,<3.0.0
loguru>=0.7.3,<0.8.0
tenacity>=9.1.2,<10.0.0
orjson>=3.10.18,<4.0.0
aiohttp>=3.12.13,<4.0.0

# --- Data & Transport ---
pandas==3.0.1
scipy==1.17.1
geopandas==1.1.3
gtfs-kit>=10.3.1,<11.0.0
arrow>=1.3.0,<2.0.0
humanize>=4.12.3,<5.0.0
faker>=37.4.0,<38.0.0
tinydb>=4.8.2,<5.0.0
cachetools>=6.0.0,<7.0.0

# --- Intelligence Artificielle & LLM ---
openai>=1.86.0,<2.0.0
faiss-cpu>=1.11.0,<2.0.0
sentence-transformers>=4.1.0,<5.0.0
nltk>=3.9.1,<4.0.0
wordcloud>=1.9.4,<2.0.0
demjson3>=3.0.6,<4.0.0

# --- LlamaIndex Ecosystem ---
llama-index>=0.12.43,<0.13.0
llama-index-embeddings-huggingface>=0.5.4,<0.6.0
llama-index-vector-stores-faiss>=0.4.0,<0.5.0
llama-index-llms-ollama>=0.6.2,<0.7.0
llama-index-llms-vllm>=0.5.1,<0.6.0
llama-index-llms-huggingface>=0.5.0,<0.6.0
llama-index-embeddings-ollama>=0.6.0,<0.7.0
llama-index-vector-stores-chroma>=0.4.2,<0.5.0

# --- Communication (MQTT) ---
asyncio-mqtt>=0.16.2,<0.17.0
paho-mqtt==1.6.1
 
# Worker
celery>=5.4.0
 
# Redis (broker + stockage)
redis>=5.0.0
 
# HTTP client (adapters LLM)
httpx>=0.27.0
 
# Templating prompts
jinja2>=3.1.4
 
# Logging structuré
structlog>=24.2.0
 
# Monitoring Celery
flower>=2.0.0

# Monitoring Prometheus
prometheus-client>=0.20.0

PyYAML>=6.0
